Move point processors to io/internal.
