+++ cartographer/io/internal/counting_points_processor.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_IO_COUNTING_POINTS_PROCESSOR_H_
-#define CARTOGRAPHER_IO_COUNTING_POINTS_PROCESSOR_H_
+#ifndef CARTOGRAPHER_IO_INTERNAL_COUNTING_POINTS_PROCESSOR_H_
+#define CARTOGRAPHER_IO_INTERNAL_COUNTING_POINTS_PROCESSOR_H_
 
 #include "cartographer/common/lua_parameter_dictionary.h"
 #include "cartographer/io/points_processor.h"
@@ -49,4 +49,4 @@ class CountingPointsProcessor : public PointsProcessor {
 }  // namespace io
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_IO_COUNTING_POINTS_PROCESSOR_H_
+#endif  // CARTOGRAPHER_IO_INTERNAL_COUNTING_POINTS_PROCESSOR_H_
