+++ cartographer/io/internal/intensity_to_color_points_processor.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_IO_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
-#define CARTOGRAPHER_IO_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
+#ifndef CARTOGRAPHER_IO_INTERNAL_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
+#define CARTOGRAPHER_IO_INTERNAL_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
 
 #include <memory>
 
@@ -61,4 +61,4 @@ class IntensityToColorPointsProcessor : public PointsProcessor {
 }  // namespace io
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_IO_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
+#endif  // CARTOGRAPHER_IO_INTERNAL_INTENSITY_TO_COLOR_POINTS_PROCESSOR_H_
