+++ cartographer/io/internal/min_max_range_filtering_points_processor.h
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-#ifndef CARTOGRAPHER_IO_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
-#define CARTOGRAPHER_IO_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
+#ifndef CARTOGRAPHER_IO_INTERNAL_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
+#define CARTOGRAPHER_IO_INTERNAL_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
 
 #include <memory>
 
@@ -55,4 +55,4 @@ class MinMaxRangeFiteringPointsProcessor : public PointsProcessor {
 }  // namespace io
 }  // namespace cartographer
 
-#endif  // CARTOGRAPHER_IO_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
+#endif  // CARTOGRAPHER_IO_INTERNAL_MIN_MAX_RANGE_FILTERING_POINTS_PROCESSOR_H_
