+++ cartographer/io/internal/ply_writing_points_processor.h
