+++ cartographer/io/points_processor_pipeline_builder.cc
@@ -17,20 +17,20 @@
 #include "cartographer/io/points_processor_pipeline_builder.h"
 
 #include "cartographer/common/make_unique.h"
-#include "cartographer/io/coloring_points_processor.h"
-#include "cartographer/io/counting_points_processor.h"
-#include "cartographer/io/fixed_ratio_sampling_points_processor.h"
-#include "cartographer/io/frame_id_filtering_points_processor.h"
-#include "cartographer/io/hybrid_grid_points_processor.h"
-#include "cartographer/io/intensity_to_color_points_processor.h"
-#include "cartographer/io/min_max_range_filtering_points_processor.h"
-#include "cartographer/io/null_points_processor.h"
-#include "cartographer/io/outlier_removing_points_processor.h"
-#include "cartographer/io/pcd_writing_points_processor.h"
-#include "cartographer/io/ply_writing_points_processor.h"
-#include "cartographer/io/probability_grid_points_processor.h"
-#include "cartographer/io/xray_points_processor.h"
-#include "cartographer/io/xyz_writing_points_processor.h"
+#include "cartographer/io/internal/coloring_points_processor.h"
+#include "cartographer/io/internal/counting_points_processor.h"
+#include "cartographer/io/internal/fixed_ratio_sampling_points_processor.h"
+#include "cartographer/io/internal/frame_id_filtering_points_processor.h"
+#include "cartographer/io/internal/hybrid_grid_points_processor.h"
+#include "cartographer/io/internal/intensity_to_color_points_processor.h"
+#include "cartographer/io/internal/min_max_range_filtering_points_processor.h"
+#include "cartographer/io/internal/null_points_processor.h"
+#include "cartographer/io/internal/outlier_removing_points_processor.h"
+#include "cartographer/io/internal/pcd_writing_points_processor.h"
+#include "cartographer/io/internal/ply_writing_points_processor.h"
+#include "cartographer/io/internal/probability_grid_points_processor.h"
+#include "cartographer/io/internal/xray_points_processor.h"
+#include "cartographer/io/internal/xyz_writing_points_processor.h"
 #include "cartographer/mapping/proto/trajectory.pb.h"
 
 namespace cartographer {
